⚠️ Moves to 16-bit stack width.
This change improved perf a tiny bit. Obviously it also decreases space-effiency drastically, which does not matter. It fixes the issue of only being able to have 255 constans per chunk (as well as 255 locals and upvalues).
Therefore we don't have to implement all the _LONG versions of the opcodes, which is a huge win. Also, this change allows for 65k+ opcodes, which I is kinda funny.