Adds universal value_to_str function. Small refactors.
This one was a huge pain in the ass. The value_to_str function seemed to have non-deterministic behaviour and crash the vm bc of violating memory accesses. In the end it was the gc who cleaned up some bytes in init_vm which led to adding the pause_gc flag. Currently, the module system is in a working, but awkward state, bc the 'export' operation is not fully implemented.