Finally, the parallel gc works, passes tests and does even improve perf in the current bench suite.
Switches from jemalloc to mimalloc, because of strange segfaults I got. In hindsight, the switch to momalloc probably wasn't necessary, because the bug was creating heap-alloc'd mem with _strudp (which I forgot to switch to je_strdup - duh) and freeing it with free (e.g. je_free).
I only realized that when I was messing around with mimalloc, so, jemalloc probably would've been fine. I - however - like mimalloc more, and, according to THEIR benches it should even outperform jemalloc.